linkerd udpate

+++ kubernetes/aks_5_service_mesh/part_1_acr/scripts/delete_aks.sh
@@ -1,2 +1,11 @@
+#!/usr/bin/env bash
+
+## Verify required commands
+command -v az > /dev/null || \
+  { echo "[ERROR]: 'az' command not not found" 1>&2; exit 1; }
+
+## Verify these variables are set
+[[ -z "$AZ_RESOURCE_GROUP" ]] && { echo 'AZ_RESOURCE_GROUP not specified. Aborting' 2>&1 ; exit 1; }
+[[ -z "$AZ_CLUSTER_NAME" ]] && { echo 'AZ_CLUSTER_NAME not specified. Aborting' 2>&1 ; exit 1; }
+
 az aks delete --resource-group ${AZ_RESOURCE_GROUP} --name ${AZ_CLUSTER_NAME}
-rm -rf ${KUBECONFIG}
